requirements.txt
python-dotenv
psycopg[binary]
argon2-cffi
pyotp
qrcode